Fix definition of wasm's long long for conversions

+++ include/rlbox_wamr_sandbox.hpp
@@ -229,7 +229,7 @@ rlbox_wamr_sandbox_thread_data* get_rlbox_wamr_sandbox_thread_data();
 class rlbox_wamr_sandbox
 {
 public:
-  using T_LongLongType = int32_t;
+  using T_LongLongType = int64_t;
   using T_LongType = int32_t;
   using T_IntType = int32_t;
   using T_PointerType = uint32_t;
